// packages/notion/src/types/enum.ts
export enum BlockType {
  PAGE = "page",
  TEXT = "text",
  BOOKMARK = "bookmark",
  BULLETED_LIST = "bulleted_list",
  NUMBERED_LIST = "numbered_list",
  HEADER = "header",
  SUB_HEADER = "sub_header",
  SUB_SUB_HEADER = "sub_sub_header",
  QUOTE = "quote",
  EQUATION = "equation",
  TO_DO = "to_do",
  TABLE_OF_CONTENTS = "table_of_contents",
  DIVIDER = "divider",
  COLUMN_LIST = "column_list",
  COLUMN = "column",
  CALLOUT = "callout",
  TOGGLE = "toggle",
  IMAGE = "image",
  EMBED = "embed",
  GIST = "gist",
  VIDEO = "video",
  FIGMA = "figma",
  TYPEFORM = "typeform",
  REPLIT = "replit",
  CODEPEN = "codepen",
  EXCALIDRAW = "excalidraw",
  TWEET = "tweet",
  MAPS = "maps",
  PDF = "pdf",
  AUDIO = "audio",
  DRIVE = "drive",
  FILE = "file",
  CODE = "code",
  COLLECTION_VIEW = "collection_view",
  COLLECTION_VIEW_PAGE = "collection_view_page",
  TRANSCLUSION_CONTAINER = "transclusion_container",
  TRANSCLUSION_REFERENCE = "transclusion_reference",
  ALIAS = "alias",
  TABLE = "table",
  TABLE_ROW = "table_row",
  EXTERNAL_OBJECT_INSTANCE = "external_object_instance",
  BREADCRUMB = "breadcrumb",
  MIRO = "miro",
  // fallback for unknown blocks
  UNKNOWN = ""
}